third_party/WebKit/LayoutTests/csspaint/registerPaint-expected.txt

CONSOLE MESSAGE: line 15: The worklet should throw an error with: "A class with name:'foo' is already registered."
CONSOLE ERROR: line 1: Uncaught [object DOMException]
CONSOLE MESSAGE: line 15: The worklet should throw an error with: "The empty string is not a valid name."
CONSOLE ERROR: line 1: Uncaught TypeError: Failed to execute 'registerPaint' on 'PaintWorkletGlobalScope': The empty string is not a valid name.
CONSOLE MESSAGE: line 15: The worklet should throw an error with: "failed!"
CONSOLE ERROR: line 1: Uncaught Error: failed!
CONSOLE MESSAGE: line 15: The worklet should throw an error with: "The value provided is neither an array, nor does it have indexed properties."
CONSOLE ERROR: line 1: Uncaught TypeError: Failed to execute 'registerPaint' on 'PaintWorkletGlobalScope': The value provided is neither an array, nor does it have indexed properties.
CONSOLE MESSAGE: line 15: The worklet should throw an error with: "The 'prototype' object on the class does not exist."
CONSOLE ERROR: line 1: Uncaught TypeError: Failed to execute 'registerPaint' on 'PaintWorkletGlobalScope': The 'prototype' object on the class does not exist.
CONSOLE MESSAGE: line 15: The worklet should throw an error with: "The 'prototype' property on the class is not an object."
CONSOLE ERROR: line 1: Uncaught TypeError: Failed to execute 'registerPaint' on 'PaintWorkletGlobalScope': The 'prototype' property on the class is not an object.
CONSOLE MESSAGE: line 15: The worklet should throw an error with: "The 'paint' function on the prototype does not exist."
CONSOLE ERROR: line 1: Uncaught TypeError: Failed to execute 'registerPaint' on 'PaintWorkletGlobalScope': The 'paint' function on the prototype does not exist.
CONSOLE MESSAGE: line 15: The worklet should throw an error with: "The 'paint' property on the prototype is not a function."
CONSOLE ERROR: line 1: Uncaught TypeError: Failed to execute 'registerPaint' on 'PaintWorkletGlobalScope': The 'paint' property on the prototype is not a function.
CONSOLE MESSAGE: line 19: The worklet should not throw an error.
CONSOLE MESSAGE: line 1: Success for 'foo9'.
CONSOLE MESSAGE: line 19: The worklet should not throw an error.
CONSOLE MESSAGE: line 1: Success for 'foo10'.
CONSOLE MESSAGE: line 19: The worklet should not throw an error.
CONSOLE MESSAGE: line 1: Success for 'foo11'.
CONSOLE MESSAGE: line 15: The worklet should throw an error with: "The 'alpha' property on the class is not a boolean."
CONSOLE ERROR: line 1: Uncaught TypeError: Failed to execute 'registerPaint' on 'PaintWorkletGlobalScope': The 'alpha' property on the class is not a boolean.
CONSOLE MESSAGE: line 19: The worklet should not throw an error.
CONSOLE MESSAGE: line 1: Success for 'foo13'.
This tests a series of PaintWorkletGlobalScope#registerPaint calls.

See the devtools console for test output.
